requirements.txt
agate==1.7.1
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dash==3.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbt-core==1.7.9
dbt-extractor==0.5.1
dbt-mysql==1.7.0
dbt-semantic-interfaces==0.4.4
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
Logbook==1.5.3
MarkupSafe==3.0.2
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msgpack==1.1.1
mysql-connector-python==9.4.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.2
parsedatetime==2.6
pathspec==0.11.2
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sqlparse==0.4.4
streamlit==1.49.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.23.0
